react (login with google)

+++ client/src/components/body/profile/Profile.js
@@ -105,15 +105,20 @@ try {
     }
     const handleDelete = async (id) => {
 try {
-    if(window.confirm("Are you sure you want to delete this account ?"))
-    {
-        setLoading(true)
-        await axios.delete(`/user/delete/${id}`,{
-            headers: {Authorization: token}
-        })
-        setLoading(false)
-        setCallback(!callback)
+
+    if(user._id !==id) {
+        if(window.confirm("Are you sure you want to delete this account ?"))
+        {
+            setLoading(true)
+            await axios.delete(`/user/delete/${id}`,{
+                headers: {Authorization: token}
+            })
+            setLoading(false)
+            setCallback(!callback)
+        }
     }
+
+  
 } catch (err) {
     setData({...data, err:err.response.data.msg ,success:''})   
 }
